[CodeCompletion] Split CodeCompletionConsumer into its own file
